risk-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
pymongo>=4.6.0
redis>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
